requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
pydantic-settings[dotenv]>=2.4
requests>=2.32
httpx>=0.27
loguru>=0.7
tenacity>=8.5

pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
xgboost>=2.1
pyarrow>=16.0
plotly>=5.24

boto3>=1.34

SQLAlchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2

streamlit>=1.36

pre-commit
black
isort
flake8
detect-secrets
pytest

